Correct inconsistencies in directory naming and request update to pages.jpg image
I noticed an inconsistency in the Directory Relative section, and while trying to figure out I also noticed that the pages.jpg file is either wrong or I'm confused. So, I'm assuming that the pages.jpg file is incorrect and that the directories under pages should be: 01.blue and 02.green and not 01.blue and 01.green.