Update docs.md
be a bit more descriptive regarding the sass compiler and how to make it run for beginners like me :) . It took a while to figure out what is going on since the little change to custom.css had no results without sass installed and running in the folder to watch the changes. Originally i thought it would be done automatically but guess not.