Clarified the second example under Field Ordering
Some new users might not notice that the example doesn't refer to a page blueprint (it defines a new field called `author`, not `header.author`). So they might try (as I did) to order a field in a page blueprint by doing `order@: title` instead of `order@: header.title`, which doesn't work.
I added a notice that helps draw attention to this.