Remove misleading umask comment (#470)
The umask is process bound and cannot be set on a per-directory basis as the comment suggested. It might however make sense to set umask in the command line *while installing Grav* to affect the subsequent invocation of **gpm**, **composer** et al.

If the advice directly below about `<?php umask(0002);` works for the whole site, then that should suffice. Otherwise, it would probably be helpful to briefly instruct the reader on how to run PHP with a different umask. If you concur, I can offer to add a small section on how to do this for systemd driven distros, which should cover most popular distros.

Please feel free to discuss or edit my commit as necessary.