Improve Modular Pages (#479)
"Modular Pages" rewritten to distinguish and clarify Modules as the building-blocks of Modular Pages, and to specify that they are not regular child-pages. Some differentiation when talking of a page or pages generally, and a Page or Pages conceptually. The same is brought to "Pages". Various linguistic improvements, formatting and tweaks on both pages.

Written as a response to getgrav/grav-plugin-admin#1174.