Changing default redirect code from 301 to 302.
A permanent redirect can not be changed remotely once it has been sent.

Permanent redirects (301 / 308) are exactly what the their descriptive
name suggest. A 301 or 308 will be indefinitely cached client-side,
sometimes even when a user clears their cache.

They should be used to keep search engine page rankings by sending a
permanent redirect on a page by page basis, not for site to site or
protocols (http to https).

For http:// to https://, the HSTS-header is much more suited, as a
client will not return to http:// for the duration that was set in the
header.

For this reason, I believe that the documentation should use temporary
redirects in examples, and if users truly believe they need 301, they
can adapt this to their needs later.

Additionally, temporary redirects can be influenced remotely and are not
that expensive on the webserver nowadays with event-based handlers.