Add extra clarification and rearrange the sections (#642)
The theme and plugins section requires further clarification. It is an important section, as any new theme developer inheriting the default theme will bump into the .php file early on in the process. Also, the newcomer will find theming easier to understand if the most important and familiar topics are introduced first, i.e., definition, templates and css, with blueprints and events as extra optional features.