Update docs.md (#912)
Hello,
I just noticed that a blueprint for a page (that is not to be included in the theme), in order to be considered in page-rendering, needs to be placed in a subfolder of user/blueprints named 'pages'. I.e. it needs to be in the folder user/blueprints/pages. If it is just in the folder user/blueprints, as the document text suggests, it seems not to be taken into account.