Update docs.md
A contraction is used in this case in most of the English speaking world to suggest something. "Let us" is more often used when meaning "we (me and someone else) want to be allowed to do something" https://grammar.collinsdictionary.com/us/english-usage/what-is-the-difference-between-let-s-and-let-us#:~:text=Let's%20decide%20what%20we%20want,something%2C%20you%20say%20let's%20not.