Update docs.md
My previous edit was incorrect. In fact, {% block content %} just needs to be implemented somewhere, not necessarily in the base.html.twig. In the case of Quark it's used in default.html.twig. It only matters that {% block content %} is found somewhere in the template inheritance hierarchy for a given page.