Update docs.md (#1113)
In the manual (non-CLI) template customization instructions, the original instructions result in a fatal error after setting-up the inherited theme. The error is: Template "@images/grav-logo.svg" is not defined in "partials/logo.html.twig" at line 7.

I traced the error to a typo or change in the original instructions on this page.  The original instructions at the prefixes section omitted using a single-quote around the prefixes and the path to the themes. Thus, on activating the theme, GRAV found nothing to inherit and failed with a fatal error. The correction is minor but critical.